Rename packages to align with existing naming convention
