+++ spring-cloud-gcp-pubsub/src/test/java/org/springframework/cloud/gcp/pubsub/integration/inbound/PubSubInboundChannelAdapterTests.java
@@ -1,0 +1,43 @@
+/*
+ *  Copyright 2017 original author or authors.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package org.springframework.cloud.gcp.pubsub.integration.inbound;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.Mock;
+import org.mockito.junit.MockitoJUnitRunner;
+
+import org.springframework.cloud.gcp.pubsub.core.PubSubOperations;
+
+/**
+ * {@link PubSubInboundChannelAdapter} unit tests.
+ *
+ * @author João André Martins
+ */
+@RunWith(MockitoJUnitRunner.class)
+public class PubSubInboundChannelAdapterTests {
+
+	@Mock
+	private PubSubOperations pubSubTemplate;
+
+	@Test(expected = IllegalArgumentException.class)
+	public void testNonNullAckMode() {
+		PubSubInboundChannelAdapter adapter = new PubSubInboundChannelAdapter(
+				this.pubSubTemplate, "testSubscription");
+		adapter.setAckMode(null);
+	}
+}
